+++ spec/integration/s3_spec.rb
@@ -12,10 +12,9 @@ describe AWS::S3 do
       end
     end
 
-
     it "should work" do
       begin
-        AWS.config( :http_handler => AWS::Http::EMHttpHandler.new(:pool_size => 20))
+        AWS.config( :http_handler => AWS::Http::EMHttpHandler.new )
         s3 = AWS::S3.new
 
         # Create bucket
@@ -23,18 +22,32 @@ describe AWS::S3 do
         bucket = s3.buckets['em_test_bucket']
         expect(bucket).to be_exists
 
-        # Write
+        # Concurrent writes
         filler = '1'*1048576
-        bucket.objects['test'].write(filler)
+
+        fibers = []
+        5.times.each do |i|
+          fiber = Fiber.new do
+            bucket.objects["test#{i}"].write(filler)
+          end
+          fiber.resume
+          fibers << fiber
+        end
+
+        # Wait until work is done
+        while fibers.detect(&:alive?)
+          EM::Synchrony.sleep(0.01)       
+        end
 
         # Streaming/Read
-        streamed = ""
-        bucket.objects["test"].read do |chunk|
+        streamed = []
+        bucket.objects["test1"].read do |chunk|
           streamed << chunk
         end
-        expect(streamed).to eql(filler)
+        expect(streamed.length).to be > 1 # make sure streaming took place
+        expect(streamed.join).to eql(filler)
       ensure
-        bucket.delete! if bucket #clean up
+        bucket.delete! if bucket # clean up
       end
     end
   end