+++ spec/spec_helper.rb
@@ -1,11 +1,9 @@
 #$LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 #$LOAD_PATH.unshift(File.dirname(__FILE__))
-require 'em_aws'
-require 'aws/core/http/em_http_handler'
+require 'em-aws'
 require 'rspec'
 require 'bundler/setup'
 require 'logger'
-require 'em-http'
 
 begin
   require 'byebug'
@@ -20,7 +18,12 @@ class StubLogger
     #we don't care
   end
 end
+
+
 AWS.config(:logger => StubLogger.new)
+
+# EM::HotTub.logger = Logger.new(STDOUT)
+# EM::HotTub.trace = true
 
 RSpec.configure do |config|